VidLP-video/requirements.txt
#METER
pytorch_lightning==1.4.0
omegaconf
torchmetrics==0.6.0
transformers==4.6.0
Pillow==8.1.0
tqdm==4.56.0
ipdb==0.13.4
numpy==1.19.5
einops==0.3.0
pyarrow
sacred==0.8.2
pandas==1.1.5
timm==0.4.12
ftfy
jsoncomment

# VideoMETER
decord==0.6.0
imageio==2.9.0
demoji==1.1.0
tslearn==0.5.2
editdistance==0.6.0
scikit-image==0.18.3
opencv-python
ffmpeg-python
pytorchvideo
pycocotools
#sudo conda install -y ffmpeg